backend/requirements.txt
# Core FastAPI and ASGI
fastapi==0.115.6
uvicorn[standard]==0.32.1
starlette==0.41.3

# Database
aiomysql==0.2.0
PyMySQL==1.1.2

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
bcrypt==4.2.1

# Data Validation & Settings
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
python-dotenv==1.0.1

# HTTP & Networking
httpx==0.28.1
requests==2.32.3
dnspython==2.6.1

# File Upload & Media
python-multipart==0.0.20
Pillow==11.0.0

# Utilities
click==8.1.7
colorama==0.4.6
email-validator==2.2.0
typing_extensions==4.12.2
annotated-types==0.7.0

# Async Support
anyio==4.7.0
greenlet==3.1.1
h11==0.14.0
sniffio==1.3.1

# Cryptography
ecdsa==0.19.1
pycryptodome==3.21.0
pyasn1==0.6.0
rsa==4.9
six==1.17.0

# LLM & AI
langchain-core==0.3.76
langchain-openai==0.3.33
openai==1.107.3
tiktoken==0.11.0

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
